+++ stoik-ktor-processor/src/main/kotlin/io/bkbn/stoik/ktor/processor/KtorProcessor.kt
@@ -1,7 +1,6 @@
 package io.bkbn.stoik.ktor.processor
 
 import com.google.devtools.ksp.processing.CodeGenerator
-import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
@@ -12,17 +11,17 @@ import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSValueArgument
 import com.google.devtools.ksp.symbol.KSVisitorVoid
 import com.google.devtools.ksp.validate
-import com.squareup.kotlinpoet.ksp.addOriginatingKSFile
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.MemberName
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.buildCodeBlock
 import com.squareup.kotlinpoet.ksp.KotlinPoetKspPreview
+import com.squareup.kotlinpoet.ksp.addOriginatingKSFile
+import com.squareup.kotlinpoet.ksp.writeTo
 import io.bkbn.stoik.ktor.core.Api
 import io.bkbn.stoik.ktor.processor.util.RouteUtils.addControlFlow
 import io.ktor.routing.Route
-import java.io.OutputStream
 
 class KtorProcessor(
   private val codeGenerator: CodeGenerator,
@@ -34,6 +33,7 @@ class KtorProcessor(
     logger.info(options.toString())
   }
 
+  @OptIn(KotlinPoetKspPreview::class)
   override fun process(resolver: Resolver): List<KSAnnotated> {
     val symbols = resolver
       .getSymbolsWithAnnotation(Api::class.qualifiedName!!)
@@ -42,15 +42,10 @@ class KtorProcessor(
     if (!symbols.iterator().hasNext()) return emptyList()
 
     symbols.forEach {
-      val file: OutputStream = codeGenerator.createNewFile(
-        dependencies = Dependencies(false),
-        packageName = "io.bkbn.stoik.generated",
-        fileName = "Apis"
-      )
-      val fb = FileSpec.builder("io.bkbn.stoik.generated", "Hi")
+      val fb = FileSpec.builder("io.bkbn.stoik.generated", "Apis")
       it.accept(Visitor(fb), Unit)
       val fs = fb.build()
-      file.write(fs.toString().toByteArray())
+      fs.writeTo(codeGenerator, false)
     }
 
     return symbols.filterNot { it.validate() }.toList()